+++ components/ui/Loader.tsx
@@ -1,0 +1,5 @@
+"use client";
+
+export default function Loader() {
+  return <span className="loader"></span>;
+}